Converted most of LoginServer to ReadPacket and std::string
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@503 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/Common/ReadPacket.h
@@ -28,7 +28,7 @@ public:
 	ReadPacket(unsigned char *buffer, size_t length);
 	
 	unsigned char getByte();
-	void skipBytes(size_t len);
+	void skipBytes(int len);
 	int getInt();
 	short getHeader(); // Basically getShort that always read at start
 	short getShort();
